requirements.txt
PyYAML==3.11
poster==0.8.1
Unirest==1.1.6
